requirements.txt
pandas>=1.5.0
matplotlib>=3.5.0
seaborn>=0.12.0
numpy>=1.23.0
pyyaml>=6.0
tqdm>=4.64.0
colorama>=0.4.6
